requirements.txt
2captcha-python==1.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==3.7.1
asyncpraw==7.8.1
asyncprawcore==2.4.0
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
fastapi==0.104.1
fastapi-cli==0.0.7
filelock==3.16.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
linkedin-api==2.3.1
lxml==5.3.0
markdown-it-py==3.0.0
markdownify==0.13.1
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
numpy==1.26.3
pandas==2.2.3
pipreqs==0.4.13
praw==7.8.1
prawcore==2.4.0
propcache==0.2.1
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jobspy==1.1.76
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
rich-toolkit==0.12.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
staffspy==0.2.20
starlette==0.27.0
tenacity==8.5.0
tldextract==5.1.3
tls-client==1.0.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
update-checker==0.18.0
urllib3==2.3.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.0.3
websocket-client==1.8.0
websockets==14.1
yarg==0.1.10
yarl==1.18.3
